separate 'bumpversion' config
Move from 'setup.cfg' to a separate file.

When command 'bumpversion' is executed (as of v0.5.3) it modifies
the files declared in its config but also the file where its own
config is. We do not want it to mess with 'setup.cfg' because it also
modifies other (unrelated) sections by applying formatting, deleting
comments, etc.