chore: Update for Django 4.2
This is a massive update to the library to incorporate all the
authentication-related changes that have been introduced to Django
since version 2.1.

The main changes are:

- Delete `.auth_backends.AbstractAuthBackend` as its purpose is now
  fulfilled by `django.contrib.auth.backends.BaseBackend`.
- Change superclass of `.auth_backends.AuthUserModelAuthBackend` from
  `AbstractAuthBackend` to `django.contrib.auth.backends.ModelBackend`,
  as the latter now works fine with custom user models that do not
  inherit from `django.contrib.auth.models.AbstractUser`.
- Delete Django management command `createsuperuser` as Django's
  built-in one now works fine.

⚠️ Breaking change:

Django app `django.contrib.auth` is now required as it is not possible
to import Django's `BaseBackend` or `ModelBackend` without it being
installed. An undesirable side effect of installing `django.contrib.auth`
is that its database migrations will be run, even though they are not
needed by this library.

📣 Announcement:

The maintainer of this library has decided to drop support for
not requiring the installation of Django app `django.contrib.auth`.
This decision was made because the work required to avoid the
installation of `django.contrib.auth` is too high, and the benefits
of doing so are not worth it. For example, the Django Admin is
tightly coupled with Django app `django.contrib.auth` and does not
work without it.