chore: Remove upper bound on Python version
This change removes the upper bound on the Python version in `pyproject.toml`.

If we become aware of any incompatibilities with future Python versions,
we can always reintroduce the upper bound at that time.
This allows users to install the package with any Python version,
including future versions, without being blocked by the upper limit.